[2015-09-23][C] Challenge #233 [Intermediate] Game of Text Life
int -> size_t for strlen return value

+++ 2015-09-23/c/challenge.c
@@ -136,7 +136,7 @@ char get_next(char *current, int row, int col, int width, int height) {
 //    printf("(%d, %d): ", row, col);
 //    puts(neighbours);
 
-    int n = strlen(neighbours);
+    size_t n = strlen(neighbours);
 
     char cell = *get(current, row , col, width);
